refactor(studio): breakout shared search functionality into separate core package
