+++ packages/sanity/src/core/search/common/tokenize.ts
