+++ packages/sanity/src/core/search/common/utils.ts
@@ -1,5 +1,5 @@
 import {ObjectSchemaType, Schema, SchemaType} from '@sanity/types'
-import {isNonNullable} from '../../../../../../util/isNonNullable'
+import {isNonNullable} from '../../util/isNonNullable'
 
 const isDocumentType = (type: SchemaType): type is ObjectSchemaType =>
   Boolean(type.type && type.type.name === 'document')
