+++ packages/sanity/src/core/search/index.ts
