+++ packages/sanity/src/core/search/search.ts
