+++ packages/sanity/src/core/search/weighted/applyWeights.test.ts
