+++ packages/sanity/src/core/search/weighted/applyWeights.ts
