+++ packages/sanity/src/core/search/weighted/index.ts
