+++ packages/sanity/src/core/search/weighted/types.ts
