+++ packages/sanity/src/core/studio/components/navbar/search/components/searchResultItem/SearchResultItem.tsx
@@ -5,9 +5,9 @@ import styled, {css} from 'styled-components'
 import {PreviewCard} from '../../../../../../components/PreviewCard'
 import {FIXME} from '../../../../../../FIXME'
 import {useSchema} from '../../../../../../hooks'
+import type {WeightedHit} from '../../../../../../search'
 import {useDocumentPresence, useDocumentPreviewStore} from '../../../../../../store'
 import {getPublishedId} from '../../../../../../util/draftUtils'
-import type {WeightedHit} from '../../search/weighted/types'
 import {useCommandList} from '../../contexts/commandList'
 import {DebugOverlay} from './DebugOverlay'
 import SearchResultItemPreview from './SearchResultItemPreview'
