+++ packages/sanity/src/core/studio/components/navbar/search/contexts/search/reducer.test.ts
@@ -1,7 +1,7 @@
 import type {CurrentUser} from '@sanity/types'
 import {act, renderHook} from '@testing-library/react'
 import {useReducer} from 'react'
-import type {SearchableType} from '../../search/weighted/types'
+import type {SearchableType} from '../../../../../../search'
 import type {RecentSearchTerms} from '../../datastores/recentSearches'
 import type {SearchOrdering} from '../../types'
 import {initialSearchState, searchReducer, SearchReducerState} from './reducer'
