+++ packages/sanity/src/core/studio/components/navbar/search/index.ts
@@ -1,3 +1,2 @@
 export * from './SearchDialog'
 export * from './SearchField'
-export * from './search'
