--- packages/sanity/src/core/studio/components/navbar/search/search/__workshop__/index.ts
@@ -1,6 +1,0 @@
-import {defineScope} from '@sanity/ui-workshop'
-import {lazy} from 'react'
-
-export default defineScope('sanity/search', 'Search', [
-  {name: 'dev', title: 'Dev', component: lazy(() => import('./DevStory'))},
-])
