+++ packages/sanity/src/core/studio/components/navbar/search/utils/hasSearchableTerms.ts
@@ -1,4 +1,4 @@
-import type {SearchTerms} from '../search/weighted/types'
+import type {SearchTerms} from '../../../../../search'
 
 export function hasSearchableTerms(terms: SearchTerms): boolean {
   return terms.query.trim() !== '' || !!terms.types.length
